refactor(delegate_to): rename with_arguments to arguments
