+++ lib/convenient_service/rspec/matchers/custom/delegate_to/entities/matcher/entities/chainings.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "chainings/sub_matchers/base"
-require_relative "chainings/sub_matchers/with_arguments"
+require_relative "chainings/sub_matchers/arguments"
 require_relative "chainings/sub_matchers/with_any_arguments"
 require_relative "chainings/sub_matchers/with_concrete_arguments"
 require_relative "chainings/sub_matchers/without_arguments"
