+++ lib/convenient_service/rspec/matchers/custom/delegate_to/entities/matcher/entities/chainings/sub_matchers/arguments/commands/apply_stub_to_track_delegations.rb
@@ -10,7 +10,7 @@ module ConvenientService
               module Entities
                 module Chainings
                   module SubMatchers
-                    class WithArguments < Chainings::SubMatchers::Base
+                    class Arguments < Chainings::SubMatchers::Base
                       module Commands
                         class ApplyStubToTrackDelegations < Support::Command
                           ##
